+++ lib/src/parameter_spec.dart
@@ -7,6 +7,8 @@ class ParameterSpec<T> implements Spec {
   T defaultValue;
   List<MetaSpec> metas = [];
   bool isSelfParameter = false;
+  bool isValue = false;
+  dynamic value;
 
   ParameterSpec.build(
     this.parameterName, {
@@ -15,8 +17,11 @@ class ParameterSpec<T> implements Spec {
     this.parameterMode = ParameterMode.normal,
     this.defaultValue,
     this.isSelfParameter = false,
+    this.isValue,
+    this.value,
   }) {
     if (metas == null) metas = [];
+    this.isValue = isValue ?? false;
   }
 
   ParameterSpec.normal(
@@ -24,12 +29,16 @@ class ParameterSpec<T> implements Spec {
     bool isSelfParameter = false,
     TypeToken type,
     List<MetaSpec> metas,
+    bool isValue,
+    dynamic value,
   }) : this.build(
           parameterName,
           type: type,
           parameterMode: ParameterMode.normal,
           metas: metas,
           isSelfParameter: isSelfParameter,
+          isValue: isValue,
+          value: value,
         );
 
   ParameterSpec.named(
@@ -38,6 +47,8 @@ class ParameterSpec<T> implements Spec {
     TypeToken type,
     T defaultValue,
     List<MetaSpec> metas,
+    bool isValue,
+    dynamic value,
   }) : this.build(
           parameterName,
           type: type,
@@ -45,6 +56,8 @@ class ParameterSpec<T> implements Spec {
           parameterMode: ParameterMode.named,
           metas: metas,
           isSelfParameter: isSelfParameter,
+          isValue: isValue,
+          value: value,
         );
 
   ParameterSpec.indexed(
@@ -53,6 +66,8 @@ class ParameterSpec<T> implements Spec {
     TypeToken type,
     T defaultValue,
     List<MetaSpec> metas,
+    bool isValue,
+    dynamic value,
   }) : this.build(
           parameterName,
           type: type,
@@ -60,17 +75,26 @@ class ParameterSpec<T> implements Spec {
           parameterMode: ParameterMode.indexed,
           metas: metas,
           isSelfParameter: isSelfParameter,
+          isValue: isValue,
+          value: value,
         );
 
   String _getType() {
     return type == null ? 'dynamic' : type.typeName;
   }
 
+  String _valueString(dynamic v) => v is String ? '"$v"' : "$v";
+
   @override
   String code({Map<String, dynamic> args = const {}}) {
-    bool withDefValue = args[KEY_WITH_DEF_VALUE] ?? false;
-    String raw = isSelfParameter ? 'this.$parameterName' : '${_getType()} $parameterName';
-    if (withDefValue && defaultValue != null) raw += '=$defaultValue';
+    String raw;
+    if (isValue) {
+      raw = parameterMode == ParameterMode.named ? "$parameterName: ${_valueString(value)}" : "${_valueString(value)}";
+    } else {
+      bool withDefValue = args[KEY_WITH_DEF_VALUE] ?? false;
+      raw = isSelfParameter ? 'this.$parameterName' : '${_getType()} $parameterName';
+      if (withDefValue && defaultValue != null) raw += '=$defaultValue';
+    }
     return raw;
   }
 }
@@ -83,8 +107,7 @@ String collectParameters(List<ParameterSpec> parameters) {
   List<String> paramsList = [];
   if (normalList.isNotEmpty) paramsList.add(normalList.map((o) => o.code()).join(", "));
   if (namedList.isNotEmpty) paramsList.add('{' + namedList.map((o) => o.code(args: {KEY_WITH_DEF_VALUE: true})).join(", ") + '}');
-  if (indexedList.isNotEmpty)
-    paramsList.add('[' + indexedList.map((o) => o.code(args: {KEY_WITH_DEF_VALUE: true})).join(", ") + ']');
+  if (indexedList.isNotEmpty) paramsList.add('[' + indexedList.map((o) => o.code(args: {KEY_WITH_DEF_VALUE: true})).join(", ") + ']');
   return paramsList.join(", ");
 }