+++ CHANGELOG.md
@@ -4,4 +4,22 @@
 
 ## 1.0.1
 
-- fixed `TypeToken`, now can use '==' operator to compare two `TypeToken` 
+- fixed `TypeToken`, now can use '==' operator to compare two `TypeToken`
+
+## 1.0.2
+
+ - remove `dart_format.dart` and replace with `package:dart_style`
+
+## 1.0.2+1
+
+- bug fixed
+
+## 1.0.2+2
+
+- bug fixed
+
+## 1.0.3
+
+- New constructor of `MetaSpec`
+    + `MetaSpec.ofInstance(String instanceName)`
+    + `MetaSpec.ofConstructor(TypeToken type, {List<ParameterSpec> parameters})`